% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_mastitis_data}
\alias{format_mastitis_data}
\title{Format the data for analysis}
\usage{
format_mastitis_data(.data, level = "lactation", nrec_min = 2)
}
\arguments{
\item{level}{Level considered for the analysis. Either 'cow' or 'lactation'}

\item{nrec_min}{Minimum number of records for a unit (cow or lactation) to be included}
}
\description{
Format the data for analysis
}
