% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM_Stan_3St_f.R
\name{hmm_3s_f}
\alias{hmm_3s_f}
\title{Three state model}
\usage{
hmm_3s_f(
  .data = data.frame(),
  level = "lactation",
  nrec_min = 2,
  Sp_a = 40,
  Sp_b = 5,
  Se1_a = 40,
  Se1_b = 5,
  Se2_a = 40,
  Se2_b = 5,
  H_to_C_mean = -1,
  H_to_C_sd = 2,
  H_to_E_mean = -1,
  H_to_E_sd = 2,
  C_to_H_mean = -3,
  C_to_H_sd = 2,
  C_to_E_mean = -100,
  C_to_E_sd = 1,
  E_to_H_mean = 10,
  E_to_H_sd = 1,
  E_to_C_mean = -100,
  E_to_C_sd = 1,
  chains = 3,
  iter_sampling = 1000
)
}
\arguments{
\item{level}{Level of interest for the analysis. Either 'cow' or 'lactation'.}

\item{nrec_min}{Minimum number of records for a unit ('cow' or 'lactation') to be included in the analysis}

\item{H_to_C_mean}{Transition from healthy to contagious - mean.}

\item{H_to_C_sd}{Transition from healthy to contagious - standard deviation.}

\item{H_to_E_mean}{Transition from healthy to environmental - mean.}

\item{H_to_E_sd}{Transition from healthy to environmental - standard deviation.}

\item{C_to_H_mean}{Transition from contagious to healthy - mean.}

\item{C_to_H_sd}{Transition from contagious to healthy - standard deviation.}

\item{C_to_E_mean}{Transition from contagious to environmental - mean.}

\item{C_to_E_sd}{Transition from contagious to environmental - standard deviation.}

\item{E_to_H_mean}{Transition from environmental to healthy - mean.}

\item{E_to_H_sd}{Transition from environmental to healthy - standard deviation.}

\item{E_to_C_mean}{Transition from environmental to contagious - mean.}

\item{E_to_C_sd}{Transition from environmental to contagious - standard deviation.}

\item{chains}{Number of MCMC chains}

\item{iter_sampling}{Number of sampling iterations}
}
\description{
Three state model
}
