% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{make_lactation}
\alias{make_lactation}
\title{Create a lactation dataset from milk recording data}
\usage{
make_lactation(.data)
}
\arguments{
\item{.data}{}
}
\description{
Create a lactation dataset from milk recording data
}
