% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scc1}
\alias{scc1}
\title{Milk recording example dataset 1}
\format{
## `scc1`
A data frame with 747 rows and 6 columns:
\describe{
  \item{rec_date}{Milk recording date}
  \item{cow_id}{Cow ID}
  \item{calv_date}{Calving date}
  \item{parity}{Cow parity}
  \item{milk}{Milk quantity in kg}
  \item{scc}{Milk somatic cell countin 1,000 cells/mL}
}
}
\usage{
scc1
}
\description{
An example milk recording dataset with 111 cows recorded in a single herd
}
\keyword{datasets}
